requirements.txt
google-generativeai==0.3.0
python-dotenv==1.0.0
streamlit==1.28.0
pandas==2.0.0
python-multipart==0.0.6
